requirements.txt
pytz~=2021.3
Django~=4.0
django-filter~=21.1
django-environ~=0.8.1
crispy-bootstrap5~=0.6
psycopg2>=2.9.1
django-crispy-forms>=1.12.0
tablib>=3.0.0
texttable>=1.6.4
django-tabination>=0.4.0
django-bootstrap-v5~=1.0.7
django-tables2~=2.4.1
discord-components~=2.1.2
discord.py>=1.3
aio-timers~=0.0.1
asgiref~=3.4.1